bruh

+++ WatchinWeebsBot/Bot.cs
@@ -211,6 +211,16 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
             //return a.Message.Author.Id.ToString() == "227462990293762049"|| a.Message.Author.Id.ToString() == "239647961502580737";
         }
 
+        bool cansendcommands = true;
+        private bool checkifcould(MessageCreateEventArgs a)
+        {
+            // Elmo 239647961502580737
+            // Klives 227462990293762049
+            // Cow 295440396006326272
+            return cansendcommands;
+            //return a.Message.Author.Id.ToString() == "227462990293762049"|| a.Message.Author.Id.ToString() == "239647961502580737";
+        }
+
         async Task Client_MessageSent(MessageCreateEventArgs e)
         {
             try
@@ -224,7 +234,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                 }
             }
             */
-                if (e.Message.Content.ToLower().Contains("hello nezuko") && e.Message.Author.Id.ToString() != "737326568271118457")
+                if (e.Message.Content.ToLower().Contains("hello nezuko") && e.Message.Author.Id.ToString() != "737326568271118457" && checkifcould(e))
                 {
                     await e.Channel.SendMessageAsync("Hello!");
                 }
@@ -235,16 +245,16 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                     await UpdateTheFunny(e);
                 }
                 */
-                if (e.Message.Content.ToLower().Contains("!commands"))
+                if (e.Message.Content.ToLower().Contains("!commands") && checkifcould(e))
                 {
                     await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Fuck you.");
                 }
-                if (e.Message.Content.ToLower().Contains("!postallofthefunny"))
+                if (e.Message.Content.ToLower().Contains("!postallofthefunny") && checkifcould(e))
                 {
                     Console.WriteLine("Someone wants all of the memes!");
                     await PostAllTheFunny(e);
                 }
-                if (e.Message.Content.ToLower().Contains("nezleave") && e.Message.Author.Id.ToString() == "227462990293762049")
+                if (e.Message.Content.ToLower().Contains("nezleave") && e.Message.Author.Id.ToString() == "227462990293762049" && checkifcould(e))
                 {
                     if(e.Message.Content.ToLower().Replace("nezleave", string.Empty) == "")
                     {
@@ -270,7 +280,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         }
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("nezping"))
+                if (e.Message.Content.ToLower().Contains("nezping") && checkifcould(e))
                 {
                     await e.Channel.SendMessageAsync("Ping: "+Client.Ping+"ms");
                 }
@@ -282,7 +292,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                     await e.Channel.SendMessageAsync("Role has been made!");
                 }
                 */
-                if (e.Message.Content.ToLower().Contains("!terrariaplayers"))
+                if (e.Message.Content.ToLower().Contains("!terrariaplayers") && checkifcould(e))
                 {
                     Console.WriteLine("Someone wants terraria players!");
                     try
@@ -298,18 +308,30 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync(ex.Message);
                     }
                 }
-                if (e.Message.Author.Id.ToString() == "453826077442179072" || e.Message.Author.Id.ToString() == "238327938859270145")
+                if (e.Message.Author.Id.ToString() == "453826077442179072" || e.Message.Author.Id.ToString() == "238327938859270145" && checkifcould(e))
                 {
                     await e.Message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":confounded:"));
                     await e.Message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":sweat_drops:"));
                 }
-                if (e.Message.Author.Id.ToString() == "295440396006326272")
+                if (e.Message.Author.Id.ToString() == "295440396006326272" && checkifcould(e))
                 {
                     await e.Message.CreateReactionAsync(DiscordEmoji.FromUnicode(Client, "🐷"));
                 }
-                if (e.Message.Content.ToLower().Contains("lmao") || e.Message.Content.ToLower().Contains("lmfao"))
+                bool enablelmao = false;
+                if (e.Message.Content.ToLower().Contains("!enablelmao")&&CheckIfCoolPerson(e) && checkifcould(e))
                 {
-                    await e.Message.Channel.SendMessageAsync("Uh-huh sure you are. You expect ME to believe that you laughed so hard your fucking ass fell off? You expect me to believe an action that only involves you exhaling, made your ASS and area that has an undefined border, mind you. Fell off? from exhaling? Yea, try harder next time. Maybe they’ll understand your communist nazi socialist propaganda.");
+                    enablelmao = true;
+                }
+                if (e.Message.Content.ToLower().Contains("!disablelmao") && CheckIfCoolPerson(e) && checkifcould(e))
+                {
+                    enablelmao = false;
+                }
+                if (e.Message.Content.ToLower().Contains("lmao") || e.Message.Content.ToLower().Contains("lmfao") && checkifcould(e))
+                {
+                    if (enablelmao)
+                    {
+                        await e.Message.Channel.SendMessageAsync("Uh-huh sure you are. You expect ME to believe that you laughed so hard your fucking ass fell off? You expect me to believe an action that only involves you exhaling, made your ASS and area that has an undefined border, mind you. Fell off? from exhaling? Yea, try harder next time. Maybe they’ll understand your communist nazi socialist propaganda.");
+                    }
                 }
                 if (e.Message.Content.ToLower().Replace(" ", string.Empty).Contains("klives"))
                 {
@@ -322,7 +344,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         Console.WriteLine(ex.Message);
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!giverole"))
+                if (e.Message.Content.ToLower().Contains("!giverole") && checkifcould(e))
                 {
                     try
                     {
@@ -353,7 +375,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync(ex.Message);
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!removerole"))
+                if (e.Message.Content.ToLower().Contains("!removerole") && checkifcould(e))
                 {
                     try
                     {
@@ -383,7 +405,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync(ex.Message);
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!postthefunny"))
+                if (e.Message.Content.ToLower().Contains("!postthefunny") && checkifcould(e))
                 {
                     await PostTheFunny(e);
                 }
@@ -400,7 +422,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!nezban"))
+                if (e.Message.Content.ToLower().Contains("!nezban") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -420,14 +442,14 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only Klives can execute this!.");
                     }
                 }
-                if (e.Message.Content.Contains("!showmeallcool"))
+                if (e.Message.Content.Contains("!showmeallcool") && checkifcould(e))
                 {
                     foreach(string item in important)
                     {
                         await e.Channel.SendMessageAsync(item);
                     }
                 }
-                if (e.Message.Content.Contains("!smsg"))
+                if (e.Message.Content.Contains("!smsg") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -443,7 +465,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("nezuko am i cool") )
+                if (e.Message.Content.ToLower().Contains("nezuko am i cool") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -455,7 +477,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
 
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("nezuko is he cool"))
+                if (e.Message.Content.ToLower().Contains("nezuko is he cool") && checkifcould(e))
                 {
                     if (important.Contains(e.Message.MentionedUsers.ElementAtOrDefault(0).Id.ToString()))
                     {
@@ -466,12 +488,12 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Channel.SendMessageAsync("No!");
                     }
                 }
-                if (e.Message.Content.Contains("!delmsg"))
+                if (e.Message.Content.Contains("!delmsg") && checkifcould(e))
                 {
                     await e.Channel.GetMessageAsync(Convert.ToUInt64(e.Message.Content.Replace("!delmsg ", string.Empty))).Result.DeleteAsync();
                     await e.Message.DeleteAsync();
                 }
-                if (e.Message.Content.ToLower().Contains("make channel nezuko"))
+                if (e.Message.Content.ToLower().Contains("make channel nezuko") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -484,7 +506,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("delete channel nezuko"))
+                if (e.Message.Content.ToLower().Contains("delete channel nezuko") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -496,7 +518,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("lock this chat") )
+                if (e.Message.Content.ToLower().Contains("lock this chat") && checkifcould(e))
                 {
                     if (e.Message.Author.Id.ToString() == "227462990293762049")
                     {
@@ -508,7 +530,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!addcool") )
+                if (e.Message.Content.ToLower().Contains("!addcool") && checkifcould(e))
                 {
                     var guy = e.Message.MentionedUsers.ElementAtOrDefault(0).Id.ToString();
                     if (CheckIfCoolPerson(e))
@@ -529,7 +551,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!removecool"))
+                if (e.Message.Content.ToLower().Contains("!removecool") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -550,7 +572,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("unlock this chat") )
+                if (e.Message.Content.ToLower().Contains("unlock this chat") && checkifcould(e))
                 {
                     if (e.Message.Author.Id.ToString() == "227462990293762049")
                     {
@@ -563,7 +585,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         }
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!trace"))
+                if (e.Message.Content.ToLower().Contains("!trace") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -571,7 +593,7 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.DeleteAsync();
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!untrace"))
+                if (e.Message.Content.ToLower().Contains("!untrace") && checkifcould(e))
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -579,10 +601,18 @@ await e.Guild.GetMemberAsync(e.Author.Id).Result.SendMessageAsync("Here are the 
                         await e.Message.DeleteAsync();
                     }
                 }
-                if (tracing && e.Message.Author.Id.ToString() == "227462990293762049" && !e.Message.Content.Contains("!trace") && !e.Message.Content.Contains("!untrace") )
+                if (tracing && e.Message.Author.Id.ToString() == "227462990293762049" && !e.Message.Content.Contains("!trace") && !e.Message.Content.Contains("!untrace") && checkifcould(e))
                 {
                     await e.Message.Channel.SendMessageAsync(e.Message.Content);
                     await e.Message.DeleteAsync();
+                }
+                if (e.Message.Content.ToLower().Contains("!nomore") && CheckIfCoolPerson(e))
+                {
+                    cansendcommands = false;
+                }
+                if (e.Message.Content.ToLower().Contains("!morepls") && CheckIfCoolPerson(e))
+                {
+                    cansendcommands = true;
                 }
             }
             catch(Exception ex)
